requirements.txt
torch
torchvision
hydra-core
numpy<2